migrate file & add math structures
